5)llm_& Prompt Engineering/01)llm-basic/gemeni_llm_stream.py
from langchain_google_genai import GoogleGenerativeAI
from dotenv import load_dotenv
import os
load_dotenv()


llm = GoogleGenerativeAI(
    model="gemini-2.5-flash",
    temperature=0.7,
    google_api_key=os.getenv("GOOGLE_API_KEY")
    )

question = input("Enter you prompt: ")


for chunk in llm.stream(question):
    print(chunk)