pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = ""
authors = [
    { name = "ITZ-HURAIRAH18", email = "muhammadabuhurairah558@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13,<4.0"

dependencies = [
    "python-dotenv>=1.2.1,<2.0.0",
    "fastapi[standard]>=0.124.0,<0.125.0",
    "uvicorn>=0.38.0,<0.39.0",
    "python-multipart>=0.0.20,<0.0.21",
    "pymongo>=4.15.5,<5.0.0",
    "alembic>=1.17.2,<2.0.0",
    "psycopg2-binary>=2.9.11,<3.0.0",
    "pyjwt>=2.10.1,<3.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "bcrypt>=4.1.3,<5.0.0",
    "numpy>=2,<2.3.0",
    "matplotlib>=3.10.8,<4.0.0",
    "pandas>=2.3.3,<3.0.0",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "tensorflow (>=2.20.0,<3.0.0)",
    "langchain-google-genai (>=4.1.2,<5.0.0)",
    "openai (>=2.14.0,<3.0.0)",
    "agents (>=1.4.0,<2.0.0)",
    "pydantic-ai (>=1.37.0,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
